Extract conversion function and add test
